A1 + A2 Refactor
